Add a way to set timeouts (#7)
It's important (especially for a node js client) to be able to properly
set timeouts, both globally per client, and per request. There could be
cases when some API endpoint could stuck or be under load and we don't
want to infinitely wait for it in this case.

So, this commit adds a way to set timeouts both for `NodeHttpClient` and
for `BrowserHttpClient`, globally per client (via constructor
arguments), and also per request - now `IRequest` has `timeout` field,
so you can set it when using `HttpClient.send()` method.

Please note this is a breaking change, since it changes the format of
the constructor args for `NodeHttpClient`.